fix: build issues with interface resolution
