+++ packages/server/internal/app/engine.go
@@ -1,0 +1,63 @@
+package app
+
+import (
+	"context"
+	"log"
+	"net/http"
+	"os/signal"
+	"syscall"
+	"time"
+
+	"github.com/cherbie/player-cms/internal/provider"
+	"github.com/gin-gonic/gin"
+)
+
+type (
+	AppEngine struct {
+		provider.Disposable
+		*gin.Engine
+	}
+)
+
+func NewAppEngine() *AppEngine {
+	return &AppEngine{nil, gin.Default()}
+}
+
+func (engine *AppEngine) Close() error {
+	return nil
+}
+
+func runServer(engine *gin.Engine) error {
+	// Create context that listens for the interrupt signal from the OS.
+	ctx, stop := signal.NotifyContext(context.Background(), syscall.SIGINT, syscall.SIGTERM)
+	defer stop()
+
+	srv := http.Server{
+		Addr:    serverConnectionString(),
+		Handler: engine,
+	}
+
+	go func() {
+		if err := srv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
+			log.Fatalf("listen: %s\n", err)
+		}
+	}()
+
+	// Listen for the interrupt signal.
+	<-ctx.Done()
+
+	// Restore default behavior on the interrupt signal and notify user of shutdown.
+	stop()
+	log.Println("shutting down gracefully, press Ctrl+C again to force")
+
+	// The context is used to inform the server it has 5 seconds to finish
+	// the request it is currently handling
+	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancel()
+	if err := srv.Shutdown(ctx); err != nil {
+		log.Fatal("Server forced to shutdown: ", err)
+	}
+
+	log.Println("Server exiting")
+	return nil
+}
