+++ packages/server/internal/app/routes.go
@@ -4,11 +4,10 @@ import (
 	"github.com/cherbie/player-cms/internal/controllers"
 	"github.com/cherbie/player-cms/internal/provider"
 	"github.com/cherbie/player-cms/internal/service"
-	"github.com/gin-gonic/gin"
 )
 
 func setupRoutes(resources *provider.ResourceManager) {
-	engine := unwrapErr(func() (any, error) { return resources.Resolve(AppEngineResourceId) }).(*gin.Engine)
+	engine := unwrapErr(func() (any, error) { return resources.Resolve(AppEngineResourceId) }).(*AppEngine)
 	playerService := unwrapErr(func() (any, error) { return resources.Resolve(PlayerServiceResourceId) }).(service.PlayerService)
 
 	healthController := controllers.NewHealthController()
