+++ packages/server/internal/app/server.go
@@ -1,19 +1,12 @@
 package app
 
 import (
-	"context"
 	"fmt"
-	"log"
-	"net/http"
-	"os/signal"
-	"syscall"
-	"time"
 
 	"github.com/cherbie/player-cms/internal/config"
 	"github.com/cherbie/player-cms/internal/crud"
 	"github.com/cherbie/player-cms/internal/provider"
 	"github.com/cherbie/player-cms/internal/service"
-	"github.com/gin-gonic/gin"
 )
 
 type App interface {
@@ -30,7 +23,7 @@ func NewApp() App {
 	resources.RegisterSingleton(DatabaseConfigServiceResourceId, defaultFactoryFunc(service.NewDatabaseConfigService()))
 	resources.RegisterSingleton(MongoDatabaseServiceResourceId, newMongoDatabaseServiceFactoryFunc())
 	resources.RegisterSingleton(PlayerServiceResourceId, newPlayerServiceFactoryFunc())
-	resources.RegisterSingleton(AppEngineResourceId, defaultFactoryFunc(gin.Default()))
+	resources.RegisterSingleton(AppEngineResourceId, defaultFactoryFunc(NewAppEngine()))
 
 	setupRoutes(resources)
 	return &appCore{resources}
@@ -41,42 +34,7 @@ func (app *appCore) Run() error {
 	if err != nil {
 		return err
 	}
-	runServer(instance.(*gin.Engine))
-	return nil
-}
-
-func runServer(engine *gin.Engine) error {
-	// Create context that listens for the interrupt signal from the OS.
-	ctx, stop := signal.NotifyContext(context.Background(), syscall.SIGINT, syscall.SIGTERM)
-	defer stop()
-
-	srv := http.Server{
-		Addr:    serverConnectionString(),
-		Handler: engine,
-	}
-
-	go func() {
-		if err := srv.ListenAndServe(); err != nil && err != http.ErrServerClosed {
-			log.Fatalf("listen: %s\n", err)
-		}
-	}()
-
-	// Listen for the interrupt signal.
-	<-ctx.Done()
-
-	// Restore default behavior on the interrupt signal and notify user of shutdown.
-	stop()
-	log.Println("shutting down gracefully, press Ctrl+C again to force")
-
-	// The context is used to inform the server it has 5 seconds to finish
-	// the request it is currently handling
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	if err := srv.Shutdown(ctx); err != nil {
-		log.Fatal("Server forced to shutdown: ", err)
-	}
-
-	log.Println("Server exiting")
+	runServer(instance.(*AppEngine).Engine)
 	return nil
 }
